extract setTheme
